Add TreemapSection typescript migration
